third_party/blink/web_tests/platform/win11-arm64/http/tests/media/mixed-range-response-expected.txt

CONSOLE ERROR: Access to audio at 'http://localhost:8000/media/resources/load-video.php?name=../../../../media/content/silence.oga&type=audio/ogg' from origin 'http://127.0.0.1:8000' has been blocked by CORS policy: No 'Access-Control-Allow-Origin' header is present on the requested resource.
CONSOLE ERROR: Access to audio at 'http://localhost:8000/media/resources/load-video.php?name=../../../../media/content/silence.oga&type=audio/ogg&cors_allow_origin=http://127.0.0.1:8000' from origin 'http://127.0.0.1:8000' has been blocked by CORS policy: No 'Access-Control-Allow-Origin' header is present on the requested resource.
CONSOLE ERROR: Access to audio at 'http://localhost:8000/resources/redirect.php?url=http%3A%2F%2Flocalhost%3A8000%2Fmedia%2Fresources%2Fload-video.php%3Fname%3D..%2F..%2F..%2F..%2Fmedia%2Fcontent%2Fsilence.oga%26type%3Daudio%2Fogg%26cors_allow_origin%3Dhttp%3A%2F%2F127.0.0.1%3A8000' from origin 'http://127.0.0.1:8000' has been blocked by CORS policy: No 'Access-Control-Allow-Origin' header is present on the requested resource.
CONSOLE ERROR: Access to audio at 'http://localhost:8000/resources/redirect.php?url=http%3A%2F%2Flocalhost%3A8000%2Fmedia%2Fresources%2Fload-video.php%3Fname%3D..%2F..%2F..%2F..%2Fmedia%2Fcontent%2Fsilence.oga%26type%3Daudio%2Fogg%26cors_allow_origin%3Dhttp%3A%2F%2F127.0.0.1%3A8000&cors_allow_origin=http://127.0.0.1:8000' from origin 'http://127.0.0.1:8000' has been blocked by CORS policy: No 'Access-Control-Allow-Origin' header is present on the requested resource.
CONSOLE ERROR: Access to audio at 'http://localhost:8000/media/resources/load-video.php?name=../../../../media/content/silence.oga&type=audio/ogg&cors_allow_origin=http://127.0.0.1:8000' (redirected from 'http://localhost:8000/media/resources/mixed-range-response.php?location=http%3A%2F%2Flocalhost%3A8000%2Fmedia%2Fresources%2Fload-video.php%3Fname%3D..%2F..%2F..%2F..%2Fmedia%2Fcontent%2Fsilence.oga%26type%3Daudio%2Fogg%26cors_allow_origin%3Dhttp%3A%2F%2F127.0.0.1%3A8000&cors_allow_origin=http://127.0.0.1:8000') from origin 'http://127.0.0.1:8000' has been blocked by CORS policy: No 'Access-Control-Allow-Origin' header is present on the requested resource.
CONSOLE ERROR: Access to audio at 'http://localhost:8000/media/resources/load-video.php?name=../../../../media/content/silence.oga&type=audio/ogg&cors_allow_origin=http://127.0.0.1:8000' (redirected from 'http://127.0.0.1:8000/media/resources/mixed-range-response.php?location=http%3A%2F%2Flocalhost%3A8000%2Fmedia%2Fresources%2Fload-video.php%3Fname%3D..%2F..%2F..%2F..%2Fmedia%2Fcontent%2Fsilence.oga%26type%3Daudio%2Fogg%26cors_allow_origin%3Dhttp%3A%2F%2F127.0.0.1%3A8000') from origin 'http://127.0.0.1:8000' has been blocked by CORS policy: No 'Access-Control-Allow-Origin' header is present on the requested resource.
CONSOLE ERROR: Access to audio at 'http://127.0.0.1:8000/media/resources/load-video.php?name=../../../../media/content/silence.oga&type=audio/ogg&cors_allow_origin=*' (redirected from 'http://localhost:8000/media/resources/mixed-range-response.php?location=http%3A%2F%2F127.0.0.1%3A8000%2Fmedia%2Fresources%2Fload-video.php%3Fname%3D..%2F..%2F..%2F..%2Fmedia%2Fcontent%2Fsilence.oga%26type%3Daudio%2Fogg%26cors_allow_origin%3D*&cors_allow_origin=http://127.0.0.1:8000') from origin 'http://127.0.0.1:8000' has been blocked by CORS policy: No 'Access-Control-Allow-Origin' header is present on the requested resource.
This is a testharness.js-based test.
FAIL Redirect from same-origin to same-origin must succeed. promise_test: Unhandled rejection with value: "error event should not be fired. url: http://127.0.0.1:8000/resources/redirect.php?url=http%3A%2F%2F127.0.0.1%3A8000%2Fmedia%2Fresources%2Fload-video.php%3Fname%3D..%2F..%2F..%2F..%2Fmedia%2Fcontent%2Fsilence.oga%26type%3Daudio%2Fogg"
FAIL Redirect from same-origin to remote-origin must succeed. promise_test: Unhandled rejection with value: "error event should not be fired. url: http://127.0.0.1:8000/resources/redirect.php?url=http%3A%2F%2Flocalhost%3A8000%2Fmedia%2Fresources%2Fload-video.php%3Fname%3D..%2F..%2F..%2F..%2Fmedia%2Fcontent%2Fsilence.oga%26type%3Daudio%2Fogg"
PASS CORS-disallowed remote-origin with crossOrigin=anonymous must fail.
FAIL CORS-allowed remote-origin with crossOrigin=anonymous must succeed. promise_test: Unhandled rejection with value: "error event should not be fired. url: http://localhost:8000/media/resources/load-video.php?name=../../../../media/content/silence.oga&type=audio/ogg&cors_allow_origin=http://127.0.0.1:8000"
FAIL Redirect from same-origin to same-origin with crossOrigin=anonymous must succeed. promise_test: Unhandled rejection with value: "error event should not be fired. url: http://127.0.0.1:8000/resources/redirect.php?url=http%3A%2F%2F127.0.0.1%3A8000%2Fmedia%2Fresources%2Fload-video.php%3Fname%3D..%2F..%2F..%2F..%2Fmedia%2Fcontent%2Fsilence.oga%26type%3Daudio%2Fogg"
PASS Redirect from same-origin to CORS-disallowed remote-origin with crossOrigin=anonymous must fail.
FAIL Redirect from same-origin to CORS-allowed remote-origin with crossOrigin=anonymous must succeed. promise_test: Unhandled rejection with value: "error event should not be fired. url: http://127.0.0.1:8000/resources/redirect.php?url=http%3A%2F%2Flocalhost%3A8000%2Fmedia%2Fresources%2Fload-video.php%3Fname%3D..%2F..%2F..%2F..%2Fmedia%2Fcontent%2Fsilence.oga%26type%3Daudio%2Fogg%26cors_allow_origin%3Dhttp%3A%2F%2F127.0.0.1%3A8000"
PASS Redirect from CORS-disallowed remote-origin to CORS-allowed remote-origin with crossOrigin=anonymous must fail.
FAIL Redirect from CORS-allowed remote-origin to CORS-allowed remote-origin with crossOrigin=anonymous must succeed. promise_test: Unhandled rejection with value: "error event should not be fired. url: http://localhost:8000/resources/redirect.php?url=http%3A%2F%2Flocalhost%3A8000%2Fmedia%2Fresources%2Fload-video.php%3Fname%3D..%2F..%2F..%2F..%2Fmedia%2Fcontent%2Fsilence.oga%26type%3Daudio%2Fogg%26cors_allow_origin%3Dhttp%3A%2F%2F127.0.0.1%3A8000&cors_allow_origin=http://127.0.0.1:8000"
FAIL Mixing same-origin responses must succeed. promise_test: Unhandled rejection with value: "error event should not be fired. url: http://127.0.0.1:8000/media/resources/mixed-range-response.php?location=http%3A%2F%2F127.0.0.1%3A8000%2Fmedia%2Fresources%2Fload-video.php%3Fname%3D..%2F..%2F..%2F..%2Fmedia%2Fcontent%2Fsilence.oga%26type%3Daudio%2Fogg"
PASS Mixing same-origin response and remote-origin response must fail.
PASS Mixing remote-origin response and same-origin response must fail.
FAIL Mixing same remote-origin responses must succeed. promise_test: Unhandled rejection with value: "error event should not be fired. url: http://localhost:8000/media/resources/mixed-range-response.php?location=http%3A%2F%2Flocalhost%3A8000%2Fmedia%2Fresources%2Fload-video.php%3Fname%3D..%2F..%2F..%2F..%2Fmedia%2Fcontent%2Fsilence.oga%26type%3Daudio%2Fogg"
FAIL Mixing same CORS-allowed remote-origin responses with crossOrigin=anonymous must succeed. promise_test: Unhandled rejection with value: "error event should not be fired. url: http://localhost:8000/media/resources/mixed-range-response.php?location=http%3A%2F%2Flocalhost%3A8000%2Fmedia%2Fresources%2Fload-video.php%3Fname%3D..%2F..%2F..%2F..%2Fmedia%2Fcontent%2Fsilence.oga%26type%3Daudio%2Fogg%26cors_allow_origin%3Dhttp%3A%2F%2F127.0.0.1%3A8000&cors_allow_origin=http://127.0.0.1:8000"
FAIL Mixing same-origin response and CORS-allowed remote-origin response with crossOrigin=anonymous must succeed. promise_test: Unhandled rejection with value: "error event should not be fired. url: http://127.0.0.1:8000/media/resources/mixed-range-response.php?location=http%3A%2F%2Flocalhost%3A8000%2Fmedia%2Fresources%2Fload-video.php%3Fname%3D..%2F..%2F..%2F..%2Fmedia%2Fcontent%2Fsilence.oga%26type%3Daudio%2Fogg%26cors_allow_origin%3Dhttp%3A%2F%2F127.0.0.1%3A8000"
FAIL Mixing CORS-allowed remote-origin response and same-origin response with crossOrigin=anonymous must succeed. promise_test: Unhandled rejection with value: "error event should not be fired. url: http://localhost:8000/media/resources/mixed-range-response.php?location=http%3A%2F%2F127.0.0.1%3A8000%2Fmedia%2Fresources%2Fload-video.php%3Fname%3D..%2F..%2F..%2F..%2Fmedia%2Fcontent%2Fsilence.oga%26type%3Daudio%2Fogg%26cors_allow_origin%3D*&cors_allow_origin=http://127.0.0.1:8000"
Harness: the test ran to completion.
